BE/requirements.txt
jaclang
jac-cloud
byllm
gitpython
uvicorn
google-generativeai>=0.3.0
tree-sitter
tree-sitter-python
tree-sitter-javascript